requirements.txt
sqlalchemy
psycopg2-binary
psycopg2
python-dotenv
numpy
pandas
